On branch installer  correct ciax-updater;

+++ utils/ciax-updater.sh
@@ -1,6 +1,11 @@
 #!/bin/bash
-cd
-[ -e ciax-xml.tgz ] && rm ciax-xml.tgz
-wget http://ciax.sum.naoj.org/dav/ciax-xml.tgz
-tar xvzf ciax-xml.tgz
-clean_up
+dir=~/ciax-xml
+if [ -d $dir/.git ] ; then
+    cd $dir;git pull
+else
+    cd
+    [ -e ciax-xml.tgz ] && rm ciax-xml.tgz
+    wget http://ciax.sum.naoj.org/dav/ciax-xml.tgz
+    tar xvzf ciax-xml.tgz
+fi
+ciax-installer
